add index alias support (#50)
Completes [ENT-128](https://linear.app/hasura/issue/ENT-128/add-aliases-to-introspection)

This PR adds support for index aliases.
More reading on aliases in Elasticsearch: https://www.elastic.co/guide/en/elasticsearch/reference/current/aliases.html

## Implementation
Each alias is added as a separate index in `configuration.json` file. This means, the connector will treat each alias as its own index. The alias is given the same mappings as that of the original index.

## Review the PR
- The fix is in [this commit](https://github.com/hasura/ndc-elasticsearch/pull/50/commits/0e140faa1d48c95b72bedaa5b37bffe62235416b)
- The [refactor commit](https://github.com/hasura/ndc-elasticsearch/pull/50/commits/44338c720325eeac22a7e37ce145d43629f4c2bb) cleans up the fix
- Documentation is updated in [this commit](https://github.com/hasura/ndc-elasticsearch/pull/50/commits/396b3bad454a26ffdbc2abd90766fa873c96da8c)

---
**NOTE**
There are no tests for this change currently, because simple unit tests cannot properly test this change. This change requires an Elasticsearch database to be running and accessible by the connector. A test suite that deploys a test db is a WiP currently.

---
**COMMITS**

* add index alias support

* refactor

* update documentation

* update changelog