add credentials provider (#65)
## About

Completes [ENT-290](https://linear.app/hasura/issue/ENT-290/add-credentials-provider-to-connector)

This PR adds support for a credentials provider as per [this RFC](https://docs.google.com/document/d/1auGoIR7_5sPSVAAsxyKcdK_bgQZeT3Uxm0CwPRrE1yY/edit?tab=t.0)

## Changes
### Required New Env Vars Introduced
- `ELASTICSEARCH_CREDENTIALS_PROVIDER_KEY`: This is the `key` for the credentials provider
- `ELASTICSEARCH_CREDENTIALS_PROVIDER_MECHANISM`: This is the security credential that is expected from the credential provider service. Could be `api-key` or `service-token`
- `HASURA_CREDENTIALS_PROVIDER_URI`: The URI for the webhook service
All these env vars are not part of the `ddn connector init` flow. The user needs to manually add them after initializing the connector.

### When to use Credentials Provider
The Connector will use the Credentials Provider service if `HASURA_CREDENTIALS_PROVIDER_URI` env var is set and is not empty

## Tests
Basic unit tests that test the assertion of correct errors are present, but a full test would require a running elasticsearch instance and an environment where variables can be set and unset. These things are not a part of the test suite for now, but will be in the future.



---

* add credentials provider

* update changelog

* update go sdk version

* add docs

* fix spelling